+++ backend/math/src/lib.rs
@@ -8,9 +8,25 @@ mod terms;
 pub mod utils;
 mod variables;
 
+use std::fmt::Display;
+
 pub use evaluables::*;
 pub use numbers::*;
 pub use polynomials::Polynomial;
 pub use terms::Term;
 pub use variables::PolynomialVariable;
 pub use variables::VariableList;
+
+/// Allows for every type with Display to immediately be printed as Typst math output: $ num $
+///
+/// This is mostly useful for individual elements which need to math formatted on their own. When
+/// you do something like `format!("${num_1} dot {num_2}$")` you still need the `format` to parse
+/// the parameters correctly, so the equivalent function call would be
+/// `format!("{num_1} dot {num_2}").as_math()`
+pub trait MathDisplay: Display {
+    fn as_math(&self) -> String {
+        format!("${self}$")
+    }
+}
+
+impl<T: Display> MathDisplay for T {}